Input: atmel_mxt_ts - Implement CRC check for configuration data.
The configuration is stored in NVRAM on the maXTouch chip. When the device is
reset it reports a CRC of the stored configuration values. Therefore it isn't
necessary to send the configuration on each probe - we can check the CRC
matches and avoid a timeconsuming backup/reset cycle.

Signed-off-by: Nick Dyer <nick.dyer@itdev.co.uk>
(cherry picked from commit 9504719cc2077f1781b9d8ff58ca52e92fec9703)