staging: android: binder: modify binder interface for 64bit systems
The changes in this patch will fix the binder driver for use with 64bit
systems without affecting the existing 32bit interface. The changes
reflect the types being passed to and from the kernel space.

Most of the changes convert longs being used as 32bit values to ints.
This will not affect the existing 32bit binder, however without this patch
when the driver is compiled with a 64bit toolchain the stuctures transferred
between kernel and user space will be different between 32 and 64bit
systems with no need (e.g. struct binder_version).

There is also one change to the BinderDriverCommandProtocol that
modifies BC_FREE_BUFFER to:

-       BC_FREE_BUFFER = _IOW('c', 3, int),
+       BC_FREE_BUFFER = _IOW('c', 3, void *)

The change reflects the type being passed from the user space-a pointer
to a binder_buffer. This change should not affect the existing 32bit
user space since BC_FREE_BUFFER is computed as:

 #define _IOW(type,nr,size)         \
    ((type) << _IOC_TYPESHIFT) |    \
    ((nr)   << _IOC_NRSHIFT) |      \
    ((size) << _IOC_SIZESHIFT))

and for a 32bit compiler BC_FREE_BUFFER will have the same computed
value. This change will also ease our work in differentiating
BC_FREE_BUFFER from COMPAT_BC_FREE_BUFFER.

This patch stands as the base for the binder compat layer.

Signed-off-by: Serban Constantinescu <serban.constantinescu@arm.com>