compiler.h: don't use temporary variable in __compiletime_assert()
Usually, BUG_ON and friends aren't even evaluated in sparse, but recently
compiletime_assert_atomic_type() was added, and that now results in a
sparse warning every time it is used.

The reason turns out to be the temporary variable, after it sparse no
longer considers the value to be a constant, and results in a warning and
an error.  The error is the more annoying part of this as it suppresses
any further warnings in the same file, hiding other problems.

Since this is all about compile time and the condition should be
side-effect free to start with, there's no downside (apart maybe from a
slight compilation time penalty?) to just duplicating it, leaving sparse
able to evaluate it at check time, getting rid of the warning and error.

Signed-off-by: Johannes Berg <johannes.berg@intel.com>
Cc: Daniel Santos <daniel.santos@pobox.com>
Tested-by: Luciano Coelho <luciano.coelho@intel.com>
Acked-by: Peter Zijlstra <peterz@infradead.org>
Acked-by: Paul E. McKenney <paulmck@linux.vnet.ibm.com>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>